requirements.txt
django==4.2.3
tenant-schemas-celery==2.2.0
django-tenants==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-environ==0.10.0
django-redis==5.3.0
django-extensions==3.2.3
django-debug-toolbar==4.1.0
Pillow==10.0.0
ipython==8.14.0
psycopg2-binary==2.9.6
requests==2.31.0
pycryptodome==3.18.0
celery==5.3.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-filter==23.2
django-cacheops==7.0.1
django-filebrowser==4.0.2
django-ckeditor==6.6.1
django-tenant-users==1.2.0